+++ src/oauth2/types.ts
@@ -12,6 +12,7 @@ export type OAuth2Token = {
   accessTokenExpires: number,
   tokenType: 'bearer',
   userId: number,
+  clientId: number,
 };
 
 export type OAuth2Code = {